requirements.txt
streamlit==1.35.0
langchain>=0.2.0,<0.3.0
langchain-community
langchain-openai
faiss-cpu==1.12.0
sentence-transformers==3.0.0
tweepy==4.14.0  # Optional
matplotlib>=3.9.2
seaborn>=0.13.2
pandas>=2.2.2
numpy>=1.26.4
python-dotenv==1.0.1